requirements.txt
asgiref==3.7.2
Django==4.2.10
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
djangorestframework-api-key==3.0.0
psycopg2-binary==2.9.2
gunicorn==21.2.0
whitenoise==6.6.0
newrelic==10.2.0
django4-background-tasks==1.2.9
django-cors-headers==4.3.1
drf-yasg==1.21.7
model-bakery==1.18.0
pytest-cov==5.0.0
pytest-django==4.8.0
drf-yasg==1.21.7
django-admin-volt==1.0.10
django-two-factor-auth==1.17
phonenumbers==8.13.49
requests==2.30.0
geopy==2.3.0
django-debug-toolbar==4.4.6
djoser==2.1.0
celery==5.3.1
redis==5.2.1
flower==2.0.1
channels==4.0.0
daphne==4.1.2
channels_redis==4.2.1
django-simple-history==3.8.0
python-dotenv ==1.0.0